feat: add blur and liquid glass example components
Add example components demonstrating blur and liquid glass effects including:
- MainExample with blur type and amount selectors
- PracticalExamples with common UI patterns
- ComparisonExamples with blur type showcase
- LiquidGlassExample with interactive glass effects
